Improve deleting files

+++ start.sh
@@ -56,7 +56,12 @@ upsert_remote_alias () {
 }
 
 delete_process_lock_files () {
-  find ./.git -type f -iname '*.lock' -delete
+  ARE_PROCEE_LOCK_FILES=$(find ./.git -type f -iname '*.lock')
+  if [ -n "$ARE_PROCEE_LOCK_FILES" ]; then
+    echo Deleting process lock files:
+    echo $ARE_PROCEE_LOCK_FILES
+    find ./.git -type f -iname '*.lock' -delete
+  fi
 }
 
 trap exit_trap EXIT
